+++ lib/src/key_pair/p256_key_pair.dart
@@ -7,8 +7,6 @@ import 'package:elliptic/elliptic.dart' as ec;
 import 'package:pointycastle/api.dart' as pc;
 
 import '../digest_utils.dart';
-import '../exceptions/ssi_exception.dart';
-import '../exceptions/ssi_exception_type.dart';
 import '../types.dart';
 import '../utility.dart';
 import './_ecdh_utils.dart' as ecdh_utils;
@@ -21,7 +19,7 @@ import 'public_key.dart';
 /// This key pair supports signing and verifying data using the
 /// `ecdsa_p256_sha256` signature scheme. It also supports Elliptic Curve
 /// Diffie-Hellman (ECDH) key agreement.
-class P256KeyPair implements KeyPair {
+class P256KeyPair extends KeyPair {
   static final ec.EllipticCurve _p256 = ec.getP256();
   final ec.PrivateKey _privateKey;
   Uint8List? _publicKeyBytes;
@@ -70,18 +68,8 @@ class P256KeyPair implements KeyPair {
   }
 
   @override
-  Future<Uint8List> sign(
-    Uint8List data, {
-    SignatureScheme? signatureScheme,
-  }) async {
-    signatureScheme ??= SignatureScheme.ecdsa_p256_sha256;
-    if (signatureScheme != SignatureScheme.ecdsa_p256_sha256) {
-      throw SsiException(
-        message:
-            'Unsupported signature scheme. Currently only ecdsa_p256_sha256 is supported with p256',
-        code: SsiExceptionType.unsupportedSignatureScheme.code,
-      );
-    }
+  Future<Uint8List> internalSign(
+      Uint8List data, SignatureScheme signatureScheme) async {
     final digest = DigestUtils.getDigest(
       data,
       hashingAlgorithm: signatureScheme.hashingAlgorithm,
@@ -91,19 +79,8 @@ class P256KeyPair implements KeyPair {
   }
 
   @override
-  Future<bool> verify(
-    Uint8List data,
-    Uint8List signature, {
-    SignatureScheme? signatureScheme,
-  }) async {
-    signatureScheme ??= SignatureScheme.ecdsa_p256_sha256;
-    if (signatureScheme != SignatureScheme.ecdsa_p256_sha256) {
-      throw SsiException(
-        message:
-            'Unsupported signature scheme. Currently only ecdsa_p256_sha256 is supported with p256',
-        code: SsiExceptionType.unsupportedSignatureScheme.code,
-      );
-    }
+  Future<bool> internalVerify(Uint8List data, Uint8List signature,
+      SignatureScheme signatureScheme) async {
     final digest = DigestUtils.getDigest(
       data,
       hashingAlgorithm: signatureScheme.hashingAlgorithm,
@@ -114,8 +91,8 @@ class P256KeyPair implements KeyPair {
   }
 
   @override
-  List<SignatureScheme> get supportedSignatureSchemes =>
-      [SignatureScheme.ecdsa_p256_sha256];
+  SignatureScheme get defaultSignatureScheme =>
+      SignatureScheme.ecdsa_p256_sha256;
 
   @override
   Future<Uint8List> encrypt(Uint8List data, {Uint8List? publicKey}) async {